+++ sync/models/__init__.py
@@ -10,6 +10,7 @@ from . import sync_project_context
 from . import sync_project_demo
 from . import sync_task
 from . import sync_job
+from . import sync_data
 from . import ir_logging
 from . import ir_actions
 from . import ir_attachment