+++ sync/tests/__init__.py
@@ -1,5 +1,6 @@
 # License MIT (https://opensource.org/licenses/MIT).
 
+from . import test_property
 from . import test_links
 from . import test_trigger_db
 from . import test_default_value